Adding screen to view the publish queue (#42)
Had to refactor the initialization of the html exporter to accommodate for some extra information that was needed

+++ apps/server/src/app/export.service.ts
@@ -7,14 +7,22 @@ export class ExportService {
 
   public cancel(packageId: string) {
     const exporter = this.exports.find(e => e.packageId === packageId);
+
+    if (!exporter) return;
+
     const index = this.exports.indexOf(exporter);
 
     if (index >= 0) {
       this.exports.splice(index, 1);
-      exporter.igPublisherProcess.kill("SIGKILL");
+
+      if (exporter.igPublisherProcess) {
+        exporter.igPublisherProcess.kill("SIGKILL");
+      }
+
       exporter.sendSocketMessage('progress', 'You have been removed from the queue');
       return true;
     }
+
     return false;
   }
 }
